# Conflicts:
#	src/main/java/com/example/waterbillingsystem/Billing/ConsumerController.java

+++ src/main/java/com/example/waterbillingsystem/Billing/ConsumerService.java
@@ -46,7 +46,7 @@ public class ConsumerService {
 
     public void delete(Integer id) throws ConsumerNotFoundException {
 
-        Long count = repo.countById(id);
+        Integer count = repo.countById(id);
         if (count == 0 || count == null) {
             throw new ConsumerNotFoundException("Could not find user with ID: " + id);
         }
